Convert WebSocket client to TS and to a more conventional class
